Moved fields outside of expressions sub-package
