+++ monggregate/expressions/expressions.py
@@ -16,7 +16,7 @@ from monggregate.base import BaseModel, pyd
 
 # Local imports
 # ----------------------------
-from monggregate.expressions.fields import FieldPath, Variable
+from monggregate.fields import FieldPath, Variable
 from monggregate.expressions.content import Content
 from monggregate.operators.accumulators import(
     Avg,
