+++ test/test_expressions.py
@@ -5,7 +5,7 @@
 import pytest
 import pydantic
 from monggregate.base import pyd
-from monggregate.expressions.fields import FieldName
+from monggregate.fields import FieldName
